[8.x] Add `validated` function in `Request`
