+++ src/Illuminate/Foundation/Http/FormRequest.php
@@ -53,78 +53,19 @@ class FormRequest extends Request implements ValidatesWhenResolved
     protected $redirectAction;
 
     /**
-     * The key to be used for the view error bag.
-     *
-     * @var string
-     */
-    protected $errorBag = 'default';
-
-    /**
-     * Indicates whether validation should stop after the first rule failure.
-     *
-     * @var bool
-     */
-    protected $stopOnFirstFailure = false;
-
-    /**
-     * The validator instance.
-     *
-     * @var \Illuminate\Contracts\Validation\Validator
-     */
-    protected $validator;
-
-    /**
-     * Get the validator instance for the request.
-     *
-     * @return \Illuminate\Contracts\Validation\Validator
-     */
-    protected function getValidatorInstance()
-    {
-        if ($this->validator) {
-            return $this->validator;
-        }
-
-        $factory = $this->container->make(ValidationFactory::class);
-
-        if (method_exists($this, 'validator')) {
-            $validator = $this->container->call([$this, 'validator'], compact('factory'));
-        } else {
-            $validator = $this->createDefaultValidator($factory);
-        }
-
-        if (method_exists($this, 'withValidator')) {
-            $this->withValidator($validator);
-        }
-
-        $this->setValidator($validator);
-
-        return $this->validator;
-    }
-
-    /**
      * Create the default validator instance.
      *
-     * @param  \Illuminate\Contracts\Validation\Factory  $factory
      * @return \Illuminate\Contracts\Validation\Validator
      */
-    protected function createDefaultValidator(ValidationFactory $factory)
+    public function validator()
     {
-        return $factory->make(
-            $this->validationData(), $this->container->call([$this, 'rules']),
-            $this->messages(), $this->attributes()
-        )->stopOnFirstFailure($this->stopOnFirstFailure);
+        return $this->validate(
+            $this->container->call([$this, 'rules']),
+            $this->messages(),
+            $this->attributes(),
+        );
     }
-
-    /**
-     * Get data to be validated from the request.
-     *
-     * @return array
-     */
-    public function validationData()
-    {
-        return $this->all();
-    }
-
+    
     /**
      * Handle a failed validation attempt.
      *
@@ -191,29 +132,6 @@ class FormRequest extends Request implements ValidatesWhenResolved
     }
 
     /**
-     * Get a validated input container for the validated input.
-     *
-     * @param  array|null  $keys
-     * @return \Illuminate\Support\ValidatedInput|array
-     */
-    public function safe(array $keys = null)
-    {
-        return is_array($keys)
-                    ? $this->validator->safe()->only($keys)
-                    : $this->validator->safe();
-    }
-
-    /**
-     * Get the validated data from the request.
-     *
-     * @return array
-     */
-    public function validated()
-    {
-        return $this->validator->validated();
-    }
-
-    /**
      * Get custom messages for validator errors.
      *
      * @return array
@@ -231,19 +149,6 @@ class FormRequest extends Request implements ValidatesWhenResolved
     public function attributes()
     {
         return [];
-    }
-
-    /**
-     * Set the Validator instance.
-     *
-     * @param  \Illuminate\Contracts\Validation\Validator  $validator
-     * @return $this
-     */
-    public function setValidator(Validator $validator)
-    {
-        $this->validator = $validator;
-
-        return $this;
     }
 
     /**
