+++ src/Illuminate/Http/Request.php
@@ -11,6 +11,9 @@ use Illuminate\Support\Traits\Macroable;
 use RuntimeException;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request as SymfonyRequest;
+use Illuminate\Contracts\Validation\Validator;
+use Illuminate\Contracts\Validation\Factory as ValidationFactory;
+use Illuminate\Validation\ValidationException;
 
 /**
  * @method array validate(array $rules, ...$params)
@@ -23,6 +26,27 @@ class Request extends SymfonyRequest implements Arrayable, ArrayAccess
         Concerns\InteractsWithFlashData,
         Concerns\InteractsWithInput,
         Macroable;
+
+    /**
+     * The key to be used for the view error bag.
+     *
+     * @var string
+     */
+    protected $errorBag = 'default';
+
+    /**
+     * Indicates whether validation should stop after the first rule failure.
+     *
+     * @var bool
+     */
+    protected $stopOnFirstFailure = false;
+
+    /**
+     * The validator instance.
+     *
+     * @var \Illuminate\Contracts\Validation\Validator
+     */
+    protected $validator;
 
     /**
      * The decoded JSON content for the request.
@@ -443,6 +467,95 @@ class Request extends SymfonyRequest implements Arrayable, ArrayAccess
         $request->setRouteResolver($from->getRouteResolver());
 
         return $request;
+    }
+
+    /**
+     * Create the default validator instance.
+     *
+     * @param  \Illuminate\Contracts\Validation\Factory  $factory
+     * @param  array   $rules
+     * @param  mixed   ...$params
+     * 
+     * @return \Illuminate\Contracts\Validation\Validator
+     */
+    protected function createDefaultValidator(ValidationFactory  $factory, array $rules, ...$params)
+    {
+        return $factory->make(
+            $this->validationData(),
+            $rules,
+            ...$params
+        )->stopOnFirstFailure($this->stopOnFirstFailure);
+    }
+
+    /**
+     * Make validate.
+     *
+     * @param  string  $errorBag
+     * @param  array   $rules
+     * @param  mixed   ...$params
+     * 
+     * @throws \Illuminate\Validation\ValidationException
+     */
+    public function validate(array $rules, ...$params)
+    {
+        if ($this->validator) {
+            return $this->validator;
+        }
+
+        $factory = app(ValidationFactory::class);
+
+        $validator = $this->createDefaultValidator($factory, $rules, ...$params);
+
+        if (method_exists($this, 'withValidator')) {
+            $this->withValidator($validator);
+        }
+
+        $this->setValidator($validator);
+
+        return $this->validator;
+    }
+
+    /**
+     * Make validate with bag.
+     *
+     * @param  string  $errorBag
+     * @param  array   $rules
+     * @param  mixed   ...$params
+     * 
+     * @throws \Illuminate\Validation\ValidationException
+     */
+    public function validateWithBag(string $errorBag, array $rules, ...$params)
+    {
+        try {
+            return $this->validate($rules, ...$params);
+        } catch (ValidationException $e) {
+            $e->errorBag = $errorBag;
+
+            throw $e;
+        }
+    }
+
+    /**
+     * Set the Validator instance.
+     *
+     * @param  \Illuminate\Contracts\Validation\Validator  $validator
+     * @return $this
+     */
+    public function setValidator(Validator $validator)
+    {
+        $this->validator = $validator;
+
+        return $this;
+    }
+
+    /**
+     * Get data to be validated from the request.
+     *
+     * @return array
+     */
+    public function validationData()
+    {
+        return $this->all();
     }
 
     /**
